Polish organisation structure UI
