+++ src/client/prisma/schema.prisma
@@ -59,6 +59,7 @@ model User {
   deletedApiKeys         APIKeys[]            @relation("deletedByUser")
   opengrounds            OpenGround[]
   organisations          OrganisationUser[]
+  projectMemberships     ProjectUser[]
 
   @@map("users")
 }
@@ -101,6 +102,7 @@ model Project {
   updatedAt       DateTime                  @updatedAt
   databaseConfigs DatabaseConfig[]
   currentForUsers OrganisationUser[]        @relation("CurrentOrganisationProject")
+  members         ProjectUser[]
 
   @@unique([organisationId, slug])
   @@map("projects")
@@ -116,11 +118,27 @@ model OrganisationUser {
   isCurrent      Boolean      @default(false)
   currentProjectId String?    @map("current_project_id")
   currentProject Project?     @relation("CurrentOrganisationProject", fields: [currentProjectId], references: [id], onDelete: SetNull)
+  projects        ProjectUser[]
   createdAt      DateTime     @default(now())
   updatedAt      DateTime     @updatedAt
 
   @@unique([organisationId, userId])
   @@map("organisation_users")
+}
+
+model ProjectUser {
+  id                   String           @id @default(cuid())
+  projectId            String           @map("project_id")
+  project              Project          @relation(fields: [projectId], references: [id], onDelete: Cascade)
+  organisationUserId   String           @map("organisation_user_id")
+  organisationUser     OrganisationUser @relation(fields: [organisationUserId], references: [id], onDelete: Cascade)
+  userId               String           @map("user_id")
+  user                 User             @relation(fields: [userId], references: [id], onDelete: Cascade)
+  createdAt            DateTime         @default(now())
+
+  @@unique([projectId, organisationUserId])
+  @@index([userId])
+  @@map("project_users")
 }
 
 model OrganisationInvitedUser {
@@ -152,7 +170,7 @@ model DatabaseConfig {
   createdByUserId String                      @map("user_id")
   createdByUser   User                        @relation(fields: [createdByUserId], references: [id])
   projectId       String?                     @map("project_id")
-  project         Project?                    @relation(fields: [projectId], references: [id])
+  project         Project?                    @relation(fields: [projectId], references: [id], onDelete: Cascade)
   dbInvitedUsers  DatabaseConfigInvitedUser[]
   apiKeys         APIKeys[]
   opengrounds     OpenGround[]
@@ -165,7 +183,7 @@ model DatabaseConfigUser {
   databaseConfigId String
   userId           String          @map("user_id")
   user             User            @relation(fields: [userId], references: [id])
-  databaseConfig   DatabaseConfig? @relation(fields: [databaseConfigId], references: [id])
+  databaseConfig   DatabaseConfig? @relation(fields: [databaseConfigId], references: [id], onDelete: Cascade)
   isCurrent        Boolean         @default(false)
   canEdit          Boolean         @default(false)
   canShare         Boolean         @default(false)
