+++ src/client/src/app/(playground)/home/page.tsx
@@ -1,12 +1,27 @@
 "use client";
+
 import Filter from "@/components/(playground)/filter";
 import { useEffect, useMemo, useState } from "react";
 import useFetchWrapper from "@/utils/hooks/useFetchWrapper";
-import { useSearchParams } from "next/navigation";
-import Dashboard, { DashboardConfig } from "../../../components/(playground)/manage-dashboard/board-creator";
+import { useRouter, useSearchParams } from "next/navigation";
+import Dashboard, {
+	DashboardConfig,
+} from "../../../components/(playground)/manage-dashboard/board-creator";
 import { useRootStore } from "@/store";
 import { getFilterDetails } from "@/selectors/filter";
+import { getCurrentOrganisation } from "@/selectors/organisation";
+import {
+	getCurrentProject,
+	getProjectIsLoading,
+	getProjectList,
+} from "@/selectors/project";
+import {
+	getDatabaseConfigList,
+	getDatabaseConfigListIsLoading,
+} from "@/selectors/database-config";
 import { getFilterParamsForDashboard } from "@/helpers/client/filter";
+import { fetchProjectList } from "@/helpers/client/project";
+import { fetchDatabaseConfigList } from "@/helpers/client/database-config";
 import Loader from "@/components/common/loader";
 import { usePageHeader } from "@/selectors/page";
 import { usePostHog } from "posthog-js/react";
@@ -17,7 +32,14 @@ import DescriptionTooltip from "@/components/common/description-tooltip";
 import { BookText } from "lucide-react";
 
 export default function DashboardPage() {
+	const router = useRouter();
 	const filter = useRootStore(getFilterDetails);
+	const currentOrg = useRootStore(getCurrentOrganisation);
+	const projects = useRootStore(getProjectList);
+	const currentProject = useRootStore(getCurrentProject);
+	const isProjectLoading = useRootStore(getProjectIsLoading);
+	const databaseConfigs = useRootStore(getDatabaseConfigList);
+	const isDatabaseConfigLoading = useRootStore(getDatabaseConfigListIsLoading);
 	const { fireRequest, isLoading } = useFetchWrapper();
 	const { fireRequest: fireRunQuery } = useFetchWrapper();
 	const [initialConfig, setInitialConfig] = useState<
@@ -27,8 +49,34 @@ export default function DashboardPage() {
 	const dashboardId = searchParams.get("dashboardId");
 	const { setHeader } = usePageHeader();
 	const posthog = usePostHog();
+	const hasProject = Boolean(currentProject?.id && (projects?.length || 0) > 0);
+	const hasDbConfig = Boolean(databaseConfigs?.length);
+	const isSetupLoading =
+		isProjectLoading ||
+		isDatabaseConfigLoading ||
+		projects === undefined ||
+		(hasProject && databaseConfigs === undefined);
 
 	useEffect(() => {
+		if (currentOrg?.id) {
+			fetchProjectList(currentOrg.id);
+		}
+	}, [currentOrg?.id]);
+
+	useEffect(() => {
+		if (currentProject?.id) {
+			fetchDatabaseConfigList(() => {});
+		}
+	}, [currentProject?.id]);
+
+	useEffect(() => {
+		if (!isSetupLoading && (!currentOrg?.id || !hasProject || !hasDbConfig)) {
+			router.replace("/onboarding");
+		}
+	}, [currentOrg?.id, hasDbConfig, hasProject, isSetupLoading, router]);
+
+	useEffect(() => {
+		if (isSetupLoading || !hasProject || !hasDbConfig) return;
 		const fetchBoardLayout = async () => {
 			try {
 				setInitialConfig(undefined);
@@ -67,7 +115,7 @@ export default function DashboardPage() {
 		};
 
 		fetchBoardLayout();
-	}, [dashboardId, fireRequest]);
+	}, [dashboardId, fireRequest, hasDbConfig, hasProject, isSetupLoading]);
 
 	const runFilters = useMemo(() => {
 		return getFilterParamsForDashboard({
@@ -82,7 +130,6 @@ export default function DashboardPage() {
 		const data = await fireRunQuery({
 			requestType: "POST",
 			url: "/api/manage-dashboard/query/run",
-			// Add params of the dashboard & add a check for edit mode to send query or not through body
 			body: JSON.stringify({
 				widgetId,
 				filter: runFilters,
@@ -93,22 +140,32 @@ export default function DashboardPage() {
 		return data.response;
 	};
 
+	if (isSetupLoading || !hasProject || !hasDbConfig) {
+		return (
+			<div className="flex h-full w-full items-center justify-center">
+				<Loader />
+			</div>
+		);
+	}
+
 	return (
 		<>
 			<div className="flex w-full items-center justify-between mb-4 gap-4">
 				{initialConfig?.description && (
-					<DescriptionTooltip description={initialConfig.description} className="ml-2 h-4 w-4" icon={<BookText className="text-stone-500 cursor-pointer" />} />
+					<DescriptionTooltip
+						description={initialConfig.description}
+						className="ml-2 h-4 w-4"
+						icon={<BookText className="text-stone-500 cursor-pointer" />}
+					/>
 				)}
 				<Filter />
 				<BoardList dashboardId={dashboardId} />
 			</div>
-			{
-				isLoading ? (
-					<div className="flex items-center justify-center h-full w-full">
-						<Loader />
-					</div>
-				) : null
-			}
+			{isLoading ? (
+				<div className="flex items-center justify-center h-full w-full">
+					<Loader />
+				</div>
+			) : null}
 			{initialConfig ? (
 				<Dashboard
 					className="h-100 overflow-y-auto"
