+++ src/client/src/app/(playground)/organisation/page.tsx
@@ -1,0 +1,5 @@
+import OrganisationPage from "@/components/(playground)/organisation/organisation-page";
+
+export default function OrganisationRoutePage() {
+	return <OrganisationPage />;
+}
