+++ src/client/src/app/(playground)/organisation/project/[projectId]/page.tsx
@@ -1,0 +1,9 @@
+import OrganisationProjectPage from "@/components/(playground)/organisation/project-page";
+
+export default function ProjectRoutePage({
+	params,
+}: {
+	params: { projectId: string };
+}) {
+	return <OrganisationProjectPage projectId={params.projectId} />;
+}
