+++ src/client/src/features/organisation.ts
@@ -1,0 +1,3 @@
+export async function onOrganisationChanged(_organisationId: string) {
+	return;
+}
