+++ src/client/src/store/database-config.ts
@@ -2,11 +2,15 @@
 import { DatabaseConfigStore } from "@/types/store/database-config";
 import { lens } from "@dhmk/zustand-lens";
 
+const initialState = {
+	ping: { status: "pending" as const },
+	isLoading: false,
+	list: undefined,
+};
 
 export const databaseConfigStoreSlice: DatabaseConfigStore = lens(
 	(setStore, getStore) => ({
-		ping: { status: "pending" },
-		isLoading: false,
+		...initialState,
 		setPing: ({ error, status }) =>
 			setStore(() => ({
 				...getStore(),
@@ -25,6 +29,10 @@ export const databaseConfigStoreSlice: DatabaseConfigStore = lens(
 			setStore(() => ({
 				...getStore(),
 				isLoading: !!isLoading,
+			})),
+		reset: () =>
+			setStore(() => ({
+				...initialState,
 			})),
 	})
 );
